fix and beautify TabList
