+++ src/renderer/components/TabList/styles.module.scss
@@ -1,23 +1,27 @@
-.tablist-header {
+.tablistHeader {
   display: flex;
   flex-direction: row;
   justify-content: start;
-  flex-grow: 3 1;
+  flex-grow: (3, 1);
 }
 
-.tablist-header-ul {
+.tablistHeaderUl {
   display: flex;
   flex-direction: row;
   flex-grow: 1;
   justify-content: start;
 }
 
-.tab-header {
+.tabHeader {
   outline: none;
   width: 100%;
-  height: 0.35rem;
-  box-shadow: 0 2.45rem 0 0;
-  .focused {
-    box-shadow: 0 2.45rem 0 0 var(--delta-btn-tab-selected-bg);
-  }
+  height: 1.4rem;
+  padding-left: 6px;
+  border-left: 6px double transparent;
+  border-radius: 6px;
+}
+
+.focused {
+  border-left-color: var(--delta-btn-tab-selected-bg) !important;
+  border-radius: 6px;
 }
